Fixed AuthenticationController so all the logic is inside of AuthenticationService and UserService.
Fixed all the getById() so they are replaced with findById()

+++ src/main/java/com/isd/libr/service/UserServiceImpl.java
@@ -3,14 +3,20 @@ package com.isd.libr.service;
 import com.isd.libr.repo.BookActionRepository;
 import com.isd.libr.repo.UserRepository;
 import com.isd.libr.web.dto.UserDto;
+import com.isd.libr.web.dto.requests.RegisterRequest;
+import com.isd.libr.web.dto.requests.UpdatePasswordRequest;
 import com.isd.libr.web.dto.requests.UpdateUserRequest;
-import com.isd.libr.web.entity.BookAction;
 import com.isd.libr.web.entity.User;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -18,21 +24,48 @@ import java.util.stream.Collectors;
 class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final BookActionRepository bookActionRepository;
+    private final PasswordEncoder passwordEncoder;
 
     @Override
     public List<UserDto> findAll() {
         return userRepository.findAll().stream().map(UserDto::from).collect(Collectors.toList());
     }
 
-    @Override
-    public void save(User user) {
-        userRepository.save(user);
+    public User create(RegisterRequest request) {
+        Optional<User> existingUser = userRepository.findByEmail(request.getEmail());
+        if (existingUser.isPresent()) {
+            throw new UserAlreadyExistsException(String.format("User with [%s] already exists! Consider logging in.", request.getEmail()));
+        }
+        User user = User.builder()
+                .email(request.getEmail())
+                .firstName(request.getFirstName())
+                .lastName(request.getLastName())
+                .age(request.getAge())
+                .password(passwordEncoder.encode(request.getPasswd()))
+                .role(request.getRole())
+                .phone(request.getPhone())
+                .build();
+        return userRepository.save(user);
     }
 
     @Override
-    public User getById(long id) throws UserNotFoundException {
-        User user = userRepository.getById(id);
-        if (user == null) {
+    public void updatePassword(long id, UpdatePasswordRequest request) throws SamePasswordException {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new UserNotFoundException(String.format("User with ID [%s] not found", id));
+        }
+        if (passwordEncoder.matches(request.getNewPassword(), user.get().getPassword())) {
+            throw new SamePasswordException("New and old passwords must be different");
+        }
+        String hashedPassword = passwordEncoder.encode(request.getNewPassword());
+        user.get().setPassword(hashedPassword);
+        userRepository.save(user.get());
+    }
+
+    @Override
+    public Optional<User> findById(long id) throws UserNotFoundException {
+        Optional<User> user = userRepository.findById(id);
+        if (!user.isPresent()) {
             throw new UserNotFoundException(
                     String.format("User with ID [%s] not found", id)
             );
@@ -43,14 +76,21 @@ class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteUserById(long id) {
-        User user = userRepository.getById(id);
-        bookActionRepository.deleteAllByUser(user);
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new UserNotFoundException(String.format("User with ID [%s] not found", id));
+        }
+        bookActionRepository.deleteAllByUser(user.get());
         userRepository.deleteById(id);
     }
 
     @Override
     public void updateUser(long id, UpdateUserRequest request) {
-        User user = userRepository.getById(id);
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()) {
+            throw new UserNotFoundException(String.format("User with ID [%s] not found", id));
+        }
+        User user = optionalUser.get();
         user.setEmail(request.getEmail());
         user.setAge(request.getAge());
         user.setFirstName(request.getFirstName());
@@ -58,5 +98,11 @@ class UserServiceImpl implements UserService {
         user.setPhone(request.getPhone());
         user.setRole(request.getRole());
         userRepository.save(user);
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new UsernameNotFoundException(String.format("User with email [%s] not found", email)));
     }
 }
